Add configurable index data layout and rollover frequency

+++ cmd/esmapping-generator/app/flags_test.go
@@ -20,8 +20,9 @@ func TestOptionsWithDefaultFlags(t *testing.T) {
 
 	assert.Equal(t, "", o.Mapping)
 	assert.Equal(t, uint(7), o.EsVersion)
-	assert.Equal(t, int64(5), o.Shards)
-	assert.Equal(t, int64(1), o.Replicas)
+	assert.Equal(t, 5, o.Shards)
+	assert.Equal(t, 1, o.Replicas)
+
 	assert.Equal(t, "", o.IndexPrefix)
 	assert.Equal(t, "false", o.UseILM)
 	assert.Equal(t, "jaeger-ilm-policy", o.ILMPolicyName)
@@ -44,8 +45,8 @@ func TestOptionsWithFlags(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, "jaeger-span", o.Mapping)
 	assert.Equal(t, uint(7), o.EsVersion)
-	assert.Equal(t, int64(5), o.Shards)
-	assert.Equal(t, int64(1), o.Replicas)
+	assert.Equal(t, 5, o.Shards)
+	assert.Equal(t, 1, o.Replicas)
 	assert.Equal(t, "test", o.IndexPrefix)
 	assert.Equal(t, "true", o.UseILM)
 	assert.Equal(t, "jaeger-test-policy", o.ILMPolicyName)
